trustedshops/requirements.txt
selenium
webdriver-manager
htmltestrunner-rv